drop `future` files
