+++ packages/cli/src/commands/logout/index.ts
@@ -6,8 +6,10 @@ import { logoutCommand } from './command';
 import { getFlagsSpecification } from '../../util/get-flags-specification';
 import output from '../../output-manager';
 import { LogoutTelemetryClient } from '../../util/telemetry/commands/logout';
-import { logout as future } from './future';
 import chalk from 'chalk';
+import { errorToString } from '@vercel/error-utils';
+import { getCommandName } from '../../util/pkg-name';
+import { revocationRequest, processRevocationResponse } from '../../util/oauth';
 
 export default async function logout(client: Client): Promise<number> {
   let parsedArgs = null;
@@ -54,5 +56,49 @@ export default async function logout(client: Client): Promise<number> {
     );
   }
 
-  return await future(client);
+  const { authConfig } = client;
+
+  if (!authConfig.token) {
+    output.note(
+      `Not currently logged in, so ${getCommandName('logout')} did nothing`
+    );
+    return 0;
+  }
+
+  output.spinner('Logging out…', 200);
+
+  const revocationResponse = await revocationRequest({
+    token: authConfig.token,
+  });
+
+  output.debug(
+    `'Revocation response:', ${await revocationResponse.clone().text()}`
+  );
+
+  const [revocationError] = await processRevocationResponse(revocationResponse);
+  let logoutError = false;
+  if (revocationError) {
+    output.error(revocationError.message);
+    output.debug(revocationError.cause);
+    output.error('Failed during logout');
+    logoutError = true;
+  }
+
+  try {
+    client.updateConfig({ currentTeam: undefined });
+    client.writeToConfigFile();
+
+    client.emptyAuthConfig();
+    client.writeToAuthConfigFile();
+    output.debug('Configuration has been deleted');
+
+    if (!logoutError) {
+      output.success('Logged out!');
+      return 0;
+    }
+  } catch (err: unknown) {
+    output.debug(errorToString(err));
+    output.error('Failed during logout');
+  }
+  return 1;
 }
